+++ src/includes/traits/Facades/BrandUrls.php
@@ -21,7 +21,15 @@ trait BrandUrls
      */
     public static function brandUrl(...$args)
     {
-        return $GLOBALS[static::class]->Utils->§BrandUrls->toBrand(...$args);
+        return $GLOBALS[static::class]->Utils->§BrandUrl->toBrand(...$args);
+    }
+
+    /**
+     * @since 160629 Initial release.
+     */
+    public static function brandUrlArg(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->§BrandUrl->brandArg(...$args);
     }
 
     /**
@@ -29,7 +37,15 @@ trait BrandUrls
      */
     public static function parentBrandUrl(...$args)
     {
-        return $GLOBALS[static::class]->Utils->§BrandUrls->toBrandParent(...$args);
+        return $GLOBALS[static::class]->Utils->§BrandUrl->toBrandParent(...$args);
+    }
+
+    /**
+     * @since 160629 Initial release.
+     */
+    public static function parentBrandUrlArg(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->§BrandUrl->brandParentArg(...$args);
     }
 
     /**
@@ -37,7 +53,15 @@ trait BrandUrls
      */
     public static function coreBrandUrl(...$args)
     {
-        return $GLOBALS[static::class]->Utils->§BrandUrls->toBrandCore(...$args);
+        return $GLOBALS[static::class]->Utils->§BrandUrl->toBrandCore(...$args);
+    }
+
+    /**
+     * @since 160629 Initial release.
+     */
+    public static function coreBrandUrlArg(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->§BrandUrl->brandCoreArg(...$args);
     }
 
     /**
@@ -45,7 +69,15 @@ trait BrandUrls
      */
     public static function brandApiUrl(...$args)
     {
-        return $GLOBALS[static::class]->Utils->§BrandUrls->toBrandApi(...$args);
+        return $GLOBALS[static::class]->Utils->§BrandUrl->toBrandApi(...$args);
+    }
+
+    /**
+     * @since 160629 Initial release.
+     */
+    public static function brandApiUrlArg(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->§BrandUrl->brandApiArg(...$args);
     }
 
     /**
@@ -53,7 +85,15 @@ trait BrandUrls
      */
     public static function parentBrandApiUrl(...$args)
     {
-        return $GLOBALS[static::class]->Utils->§BrandUrls->toBrandParentApi(...$args);
+        return $GLOBALS[static::class]->Utils->§BrandUrl->toBrandParentApi(...$args);
+    }
+
+    /**
+     * @since 160629 Initial release.
+     */
+    public static function parentBrandApiUrlArg(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->§BrandUrl->brandParentApiArg(...$args);
     }
 
     /**
@@ -61,7 +101,15 @@ trait BrandUrls
      */
     public static function coreBrandApiUrl(...$args)
     {
-        return $GLOBALS[static::class]->Utils->§BrandUrls->toBrandCoreApi(...$args);
+        return $GLOBALS[static::class]->Utils->§BrandUrl->toBrandCoreApi(...$args);
+    }
+
+    /**
+     * @since 160629 Initial release.
+     */
+    public static function coreBrandApiUrlArg(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->§BrandUrl->brandCoreApiArg(...$args);
     }
 
     /**
@@ -69,7 +117,15 @@ trait BrandUrls
      */
     public static function brandCdnUrl(...$args)
     {
-        return $GLOBALS[static::class]->Utils->§BrandUrls->toBrandCdn(...$args);
+        return $GLOBALS[static::class]->Utils->§BrandUrl->toBrandCdn(...$args);
+    }
+
+    /**
+     * @since 160629 Initial release.
+     */
+    public static function brandCdnUrlArg(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->§BrandUrl->brandCdnArg(...$args);
     }
 
     /**
@@ -77,7 +133,15 @@ trait BrandUrls
      */
     public static function parentBrandCdnUrl(...$args)
     {
-        return $GLOBALS[static::class]->Utils->§BrandUrls->toBrandParentCdn(...$args);
+        return $GLOBALS[static::class]->Utils->§BrandUrl->toBrandParentCdn(...$args);
+    }
+
+    /**
+     * @since 160629 Initial release.
+     */
+    public static function parentBrandCdnUrlArg(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->§BrandUrl->brandParentCdnArg(...$args);
     }
 
     /**
@@ -85,7 +149,15 @@ trait BrandUrls
      */
     public static function coreBrandCdnUrl(...$args)
     {
-        return $GLOBALS[static::class]->Utils->§BrandUrls->toBrandCoreCdn(...$args);
+        return $GLOBALS[static::class]->Utils->§BrandUrl->toBrandCoreCdn(...$args);
+    }
+
+    /**
+     * @since 160629 Initial release.
+     */
+    public static function coreBrandCdnUrlArg(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->§BrandUrl->brandCoreCdnArg(...$args);
     }
 
     /**
@@ -93,7 +165,15 @@ trait BrandUrls
      */
     public static function brandStatsUrl(...$args)
     {
-        return $GLOBALS[static::class]->Utils->§BrandUrls->toBrandStats(...$args);
+        return $GLOBALS[static::class]->Utils->§BrandUrl->toBrandStats(...$args);
+    }
+
+    /**
+     * @since 160629 Initial release.
+     */
+    public static function brandStatsUrlArg(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->§BrandUrl->brandStatsArg(...$args);
     }
 
     /**
@@ -101,7 +181,15 @@ trait BrandUrls
      */
     public static function parentBrandStatsUrl(...$args)
     {
-        return $GLOBALS[static::class]->Utils->§BrandUrls->toBrandParentStats(...$args);
+        return $GLOBALS[static::class]->Utils->§BrandUrl->toBrandParentStats(...$args);
+    }
+
+    /**
+     * @since 160629 Initial release.
+     */
+    public static function parentBrandStatsUrlArg(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->§BrandUrl->brandParentStatsArg(...$args);
     }
 
     /**
@@ -109,6 +197,14 @@ trait BrandUrls
      */
     public static function coreBrandStatsUrl(...$args)
     {
-        return $GLOBALS[static::class]->Utils->§BrandUrls->toBrandCoreStats(...$args);
+        return $GLOBALS[static::class]->Utils->§BrandUrl->toBrandCoreStats(...$args);
+    }
+
+    /**
+     * @since 160629 Initial release.
+     */
+    public static function coreBrandStatsUrlArg(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->§BrandUrl->brandCoreStatsArg(...$args);
     }
 }